base!: Simplify by moving Boost serialization support code to `BaseMQ`
Motivation for this breaking change is to drop the `Boost::serialization`
dependency from `FairRoot::Base`.

The user must now opt-in to Boost serialization support by including one of
the new class-based headers
* `<FairTimeStampBoostSerializationSupport.h>`
* `<FairHitBoostSerializationSupport.h>`
* `<FairMCPointBoostSerializationSupport.h>`
or the convenience header `<FairBoostSerializationSupport.h>`.

resolves #1260
