+++ base/event/FairTimeStamp.h
@@ -107,39 +107,4 @@ inline FairTimeStamp::FairTimeStamp(Double_t time, Double_t timeerror)
     , fTimeStampError(timeerror)
 {}
 
-#ifdef FAIRROOT_HAS_BOOST_SERIALIZATION
-
-// #include <boost/serialization/base_object.hpp>
-#include <boost/serialization/split_free.hpp>
-
-namespace boost::serialization {
-
-template<class Archive>
-void load(Archive& ar, FairTimeStamp& time, const unsigned int)
-{
-    // ar & boost::serialization::base_object<FairMultiLinkedData>(time);
-
-    Double_t timeStamp = 0.;
-    ar& timeStamp;
-    time.SetTimeStamp(timeStamp);
-
-    Double_t timeStampError = 0.;
-    ar& timeStampError;
-    time.SetTimeStampError(timeStampError);
-}
-
-template<class Archive>
-void save(Archive& ar, FairTimeStamp const& time, const unsigned int)
-{
-    // ar & boost::serialization::base_object<FairMultiLinkedData>(time);
-    ar& time.GetTimeStamp();
-    ar& time.GetTimeStampError();
-}
-
-}   // namespace boost::serialization
-
-BOOST_SERIALIZATION_SPLIT_FREE(FairTimeStamp)
-
-#endif   // FAIRROOT_HAS_BOOST_SERIALIZATION
-
 #endif   // FAIRTIMESTAMP_H
