+++ basemq/FairBoostSerializationSupport.h
@@ -1,0 +1,18 @@
+/********************************************************************************
+ *    Copyright (C) 2022 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH    *
+ *                                                                              *
+ *              This software is distributed under the terms of the             *
+ *              GNU Lesser General Public Licence (LGPL) version 3,             *
+ *                  copied verbatim in the file "LICENSE"                       *
+ ********************************************************************************/
+
+#ifndef FAIR_BOOST_SERIALIZATION_SUPPORT_H
+#define FAIR_BOOST_SERIALIZATION_SUPPORT_H
+
+// convenience header
+
+#include <FairHitBoostSerializationSupport.h>
+#include <FairMCPointBoostSerializationSupport.h>
+#include <FairTimeStampBoostSerializationSupport.h>
+
+#endif   // FAIR_BOOST_SERIALIZATION_SUPPORT_H
