+++ examples/MQ/serialization/data/MyHit.h
@@ -1,5 +1,5 @@
 /********************************************************************************
- *    Copyright (C) 2014 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH    *
+ * Copyright (C) 2014-2022 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH  *
  *                                                                              *
  *              This software is distributed under the terms of the             *
  *              GNU Lesser General Public Licence (LGPL) version 3,             *
@@ -18,6 +18,7 @@
 
 #include "FairHit.h"   // for FairHit
 
+#include <FairHitBoostSerializationSupport.h>
 #include <Rtypes.h>   // for MyHit::Class, etc
 
 namespace boost {
